Prevents a race condition + changes debug printing
A race was possible because the uuid is only known after insertion in the database.
Indeed, a quick controller could answer before the uuid is known by the proxy.
Another ID is added. This ID is composed of an ID (i.e., a uuid of a new table and hence unique thanks to OVSDB)
 and a counter for each proxy thread inserting flow requests.
This ID is thus unique and known before its insertion in the database.