Insert a DNS forwarder in the access router
In order to gain time, most of the dns proxy functionalities
are transfered near to the controller. The DNS forwarder is built
to forward the DNS requests and to wait for SR rules to be inserted
before transfering the DNS reply to the client. This last functionnality
was implemented in the DNS proxy. The rest of the DNS proxy
implementation is quite similar. The point is to reduce the
communication time with the OVSDB server.