Enables srdns to specify source prefixes to apps
This commit changes one big assumption of the code. The binding segment
does not fully represents a flow. Actually, we can (but we are not
forced) to define, for a single flow, one binding segment by source
prefix. This is useful for external destinations that can be reached
through several providers (i.e., the segment list will lead packets to
the egress router of each of these providers).