Reduce CPU consumption of transaction threads
Transaction threads were making a loop without any sleep while waiting
for transactions to send or messages from the OVSDB server.
They were consuming a lot of CPU for nothing.
This commit uses eventfd(3) to trigger the poll when a new transaction is
waiting in the queue. This poll does not need a timeout anymore and this
significantly reduces the CPU usage.