Use netlink to add/change routes in sr-routed
Spawning a new process with iproute at each insertion adds a lot of delay.
This commit uses netlink to add the mapping between the binding segment
and the list of segments. This commit uses the newly standardized way
of mapping a segment to the encapsulation of another SRH
(see http://segment-routing.org/index.php/Implementation/AdvancedConf).